# client/server/openapi.yaml
openapi: 3.0.3
info:
  title: Quiz 8 Mastery API
  version: 1.0.0
  description: |
    Professional Vue 3 interview preparation platform with comprehensive question banks and advanced analytics.

    ## Authentication
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```

    ## Rate Limiting
    All endpoints are rate-limited to prevent abuse. Limits vary by endpoint:
    - Auth endpoints: 10-20 requests per 15 minutes
    - Payment endpoints: 20-30 requests per 15 minutes
    - Question endpoints: 100 requests per minute

    ## Versioning
    The API is versioned using URL path versioning: `/api/v1/`

  contact:
    name: Hussein Tirawi
    url: https://github.com/htirawi/q8m
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.quiz-platform.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
    Error:
      type: object
      required:
        - code
        - error
        - message
      properties:
        code:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid input provided
...
    Question:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          enum: [multiple-choice, true-false, code-completion]
          example: "multiple-choice"
        content:
          type: string
          example: "What is the Vue 3 Composition API?"
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "medium"
        level:
          type: string
          enum: [junior, intermediate, senior]
          example: "intermediate"
        framework:
          type: string
          enum: [vue, react, angular, general]
          example: "vue"
        category:
          type: string
          example: "Composition API"
        points:
          type: integer
          example: 10
        options:
          type: array
          items:
            type: object
          example: [{ "text": "Option A", "isCorrect": true }]
...
paths:
...
  /questions:
    get:
      summary: Get questions
      tags:
        - Questions
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [easy, medium, hard]
          description: Filter by difficulty
        - in: query
          name: framework
          schema:
            type: string
            enum: [vue, react, angular, general]
          description: Filter by framework
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of questions to return
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  questions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Question"
                  total:
                    type: integer
                  page:
                    type: integer
        "403":
          description: Access denied (plan upgrade required)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      requiredPlan:
                        type: string
                      suggestedPlan:
                        type: string
                      upgradeUrl:
                        type: string
  /questions/submit:
    post:
      summary: Submit answer
      tags:
        - Questions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - questionId
                - selectedAnswer
              properties:
                questionId:
                  type: string
                selectedAnswer:
                  type: string
      responses:
        "200":
          description: Answer submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  isCorrect:
                    type: boolean
                  correctAnswer:
                    type: string
                  explanation:
                    type: string
...